bosminer: api: Move sources of server module to lib
